feat(mcp): implement and rename news resources
- Renamed MCP resource files to follow `{entity}.ts` convention (`news.ts`, `item.ts`, `user.ts`).
- Updated news resource to support `hn://{topic}` URI pattern with pagination.
- Updated imports in resource files to use direct paths from `../../api/types`.
- Updated integration tests to match new file names.
- Removed outdated TODO from `src/server.ts`.
- Verified changes with integration tests.
